--- src/Landing.js
@@ -1,11 +1,0 @@
-import React from "react";
-import NavBar from "./NavBar";
-
-export default function Landing() {
-  return (
-    <div className="App">
-      <NavBar />
-      {/* <Banner/>   */}
-    </div>
-  );
-}
