--- src/Property.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Property() {
-  return <div>Inside property page</div>;
-}
