+++ src/css/App.css
@@ -1,7 +1,7 @@
 .bg-myRed {
-  background-color: crimson;
+	background-color: blue;
 }
 
 .content {
-  padding: 12px;
+	padding: 12px;
 }
