--- src/css/Rent.css
@@ -1,135 +1,0 @@
-body {
-  background-color: rgba(0, 0, 0, 0.712);
-}
-.filters {
-  border-style: solid;
-  height: 75px;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  margin-left: 0px;
-  background-color: rgb(0, 0, 94);
-}
-
-.filters div {
-  margin-left: 10px;
-}
-.buy-rent {
-  /* border-style: dashed; */
-  width: 200px;
-}
-
-.buy-rent input {
-  height: 100%;
-  width: 50%;
-}
-.price-filter {
-  width: 200px;
-  height: 100%;
-  /* border-style: dashed; */
-  display: inline;
-}
-.price-filter input {
-  width: 50%;
-}
-.price-header {
-  margin-left: 50px;
-}
-
-.bed-bath {
-  /* border-style: dashed; */
-  width: 200px;
-  height: 100%;
-  display: flex;
-}
-
-.bed-bath div {
-  margin-left: 10px;
-}
-
-.bedroom input {
-  width: 100%;
-}
-.bathroom input {
-  width: 100%;
-}
-.bed-header {
-  font-size: 20px;
-}
-.bath-header {
-  font-size: 20px;
-}
-
-h4 {
-  color: white;
-}
-
-.house-result {
-  margin-top: 25px;
-  height: 150px;
-  width: 100%;
-  border-style: dashed;
-  background-color: rgb(145, 138, 138);
-  display: flex;
-}
-.house-icon {
-  height: 100%;
-  width: 150px;
-  border-style: dashed;
-}
-
-.house-icon h3 {
-  margin-top: 35%;
-  margin-left: 20%;
-}
-.abb {
-  margin-left: 40px;
-  /* border-style: dashed; */
-  width: 300px;
-}
-
-.house-price {
-  margin-left: 50px;
-  margin-top: 30px;
-}
-
-.house-price {
-  margin-left: 50px;
-  margin-top: 30px;
-}
-
-.house-price p {
-  font-size: 30px;
-}
-
-a:link {
-  text-decoration: none;
-  color: black;
-}
-
-a:visited {
-  text-decoration: none;
-  color: black;
-}
-
-.order-title {
-  justify-content: center;
-  width: 80px;
-}
-
-.order-by {
-  margin-top: 30px;
-  display: flex;
-  width: 305px;
-  height: 50px;
-  border-style: solid;
-}
-
-.orders:hover {
-  color: blue;
-}
-
-.order-by .orders {
-  justify-content: center;
-  width: 75px;
-}
