+++ src/pages/About.js
@@ -1,5 +1,10 @@
 import React from "react";
 
 export default function About() {
-  return <div>About this site</div>;
+	return (
+		<div>
+			{" "}
+			<h1>Hello.</h1>
+		</div>
+	);
 }
